update some backend functions, separate models in client
