+++ server/Serials/Model/Chapter.hs
@@ -21,7 +21,7 @@ import qualified Database.RethinkDB.NoClash as R
 import Database.RethinkDB.NoClash hiding (table)
 
 import Safe (headMay)
-import Serials.Lib.Crud
+import Serials.Model.Lib.Crud
 import Serials.Link.Link
 
 import Numeric
