+++ server/Serials/Model/Source.hs
@@ -16,7 +16,7 @@ import GHC.Generics
 import qualified Database.RethinkDB.NoClash as R
 import Database.RethinkDB.NoClash hiding (table)
 
-import Serials.Lib.Crud
+import Serials.Model.Lib.Crud
 import Serials.Model.Scan
 import Serials.Link.Import (ImportSettings)
