+++ web/app/helpers.js
@@ -1,5 +1,7 @@
 // @flow
 
+import moment from 'moment'
+
 export var updateLocalStorage = function(key, data) {
     return localStorage.setItem(key, JSON.stringify(data))
 }
@@ -7,5 +9,10 @@ export var updateLocalStorage = function(key, data) {
 
 export var getLocalStorage = function(key) {
     return JSON.parse(localStorage.getItem(key))
+}
+
+export function toDateString(str:string):string {
+  var date = moment(str)
+  return date.calendar()
 }
